model_1_training/requirements.txt
# ============================================================
# Model 1 Training Pipeline — Requirements
# Install: pip install -r requirements.txt
#
# Platform notes:
#   CUDA (Linux/Windows NVIDIA GPU):
#     pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
#   CPU-only:
#     Standard torch from pip (fp32, slower)
# ============================================================

# --- Core ML ---
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.29.0

# --- Data ---
polars>=1.0.0
pyarrow>=15.0.0
datasets>=2.18.0

# --- Training ---
scikit-learn>=1.3.0

# --- Hyperparameter Tuning ---
ray[tune]>=2.10.0
optuna>=3.5.0

# --- Explainability ---
captum>=0.7.0

# --- Bias Detection ---
fairlearn>=0.9.0

# --- Experiment Tracking ---
wandb>=0.17.0

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0

# --- Config & Logging ---
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# --- Utilities ---
numpy>=1.24.0

# --- GCP (local machine only — not needed inside Docker) ---
kfp>=2.7.0
google-cloud-aiplatform>=1.48.0
google-cloud-storage>=2.14.0

# --- Dev/test ---
pytest>=7.4.0
